feat: add client helper methods
